Update loader/loader.py with loader/loader_new.py
